+++ database/migrations/2023_12_06_181615_decision_matrix.php
@@ -19,7 +19,7 @@ return new class extends Migration
             $table->foreign('id_kriteria')->references('id')->on('kriteria')->onDelete('cascade');
             $table->unsignedBigInteger('id_alternatif');
             $table->foreign('id_alternatif')->references('id')->on('alternatif')->onDelete('cascade');
-            $table->decimal('value', 8, 2);
+            $table->double('value');
             $table->timestamps();
         });
     }
